randint.md
In this code, we first import the random module to generate a random number. The randint() function is used to generate a random integer between 1 and 10, inclusive. Then, the user is prompted to enter their guess using the input() function, and the guess is converted to an integer using the int() function.